Proactively rebuild screens when styles are rebuilt.
Outside of the init phase. This prevents screens from being run
twice - once at an old version, and once at the new - when
a style rebuild occurs.

This manifested in problems with the CTC indicator, since Character
would hang the CTC stuff on the Text object created by the old
screen version.

Fixes #811.